Automatically build sphinx projects
This adds automatic html builds for sphinx projects by creating a
python virtual environment, installing sphinx into it, and then
using that sphinx to do the build. This has the advantage of ensuring
the build is completely isolated from the system python, so we can
ensure the precise versions we want without issue.

The disadvantage of doing it this way is that it's slow since it has
to create a virtualenv on every invocation and for every projection
since projections run in parallel and a shared cache can't be used.
One potential alternative is to just check to see is sphinx and the
required dependencies are already installed, but that requires
parsing the python dependency format. If we had projection config,
we could also just make sphinx-build binary an option.