Support sphinx integration configuration
This updates the sphinx integration to have a built-in configuration
object that it pulls various settings from. When Smithy 1.41 is
released, these will be easily configurable from the
smithy-build.json configuration.

This also slightly changes the doc settings's node parsing function
so it can be discovered by Smithy's NodeMapper, which lets us use
directed codegen's automatic settings mapping, which in turn will
automatically handle passing along integration settings when Smithy
1.41 is released.

One more change will be needed when 1.41 is released: the configure
method will need to be implemented to actually properly set the
settings.