WIP: snippet generator interface
This adds and makes use of a snippet generator interface that lets
other generators provide examples to be included in the docs.

It's unfortunately more ugly than I'd like, notbably with config.
Generators will essentially have to run much of their generator to
be able to provide an accurate example, so we have to give a fake
plugin context that they can use.

The codegen director will need to be updated to make this more
workable.

This is just a proof of concept meant to show how it might work from
the doc generator side.