Refactor structured shape generation
The page generation for most shapes with members was nearly
identical, with only the member listing type being changed and maybe
a bit of text in a section being added. This refactors all those down
to a single generator that takes the listing type as a parameter, and
moves any further changes into builtin interceptors.