Add noReplace note
This adds a note to the put operation of resource if it has the
noReplace trait indicating that it can't be used to update the
resource, linking to the update operation if one is available.

This also updates the reference formatter implementation to
unwrap optionals like the literal and string formatters do.